ai-travel-assistant/requirements.txt
fastapi
uvicorn
streamlit
requests
openai
faiss-cpu
python-dotenv
numpy